move code in own file
